Домашнее задание №16
Работа с Retrofit перенесена в data-слой и была протестирована.
