--- app/src/main/java/com/example/playlistmaker/MediaActivity.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MediaActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_media)
-    }
-}
