+++ app/src/main/java/com/example/playlistmaker/SearchAdapter.kt
@@ -3,7 +3,7 @@ package com.example.playlistmaker
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.search.domain.models.Track
 
 class SearchAdapter(
     private val tracks: List<Track>,
