+++ app/src/main/java/com/example/playlistmaker/SearchHistory.kt
@@ -1,7 +1,7 @@
 package com.example.playlistmaker
 
 import android.content.SharedPreferences
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.search.domain.models.Track
 import com.google.gson.Gson
