+++ app/src/main/java/com/example/playlistmaker/creator/Creator.kt
@@ -1,18 +1,24 @@
 package com.example.playlistmaker.creator
 
+import android.content.Context
 import android.content.Intent
 import android.media.MediaPlayer
-import com.example.playlistmaker.data.repository.PlayerRepositoryImpl
-import com.example.playlistmaker.domain.impl.PlayerInteractorImpl
-import com.example.playlistmaker.domain.models.Track
-import com.example.playlistmaker.domain.repository.PlayerRepository
-import com.example.playlistmaker.domain.use_case.PlayerInteractor
+import com.example.playlistmaker.player.data.repository.PlayerRepositoryImpl
+import com.example.playlistmaker.player.domain.impl.PlayerInteractorImpl
+import com.example.playlistmaker.player.domain.repository.PlayerRepository
+import com.example.playlistmaker.player.domain.use_case.PlayerInteractor
+import com.example.playlistmaker.search.data.network.RetrofitNetworkClient
+import com.example.playlistmaker.search.data.repository.TracksRepositoryImpl
+import com.example.playlistmaker.search.domain.api.TracksInteractor
+import com.example.playlistmaker.search.domain.api.TracksRepository
+import com.example.playlistmaker.search.domain.impl.TracksInteractorImpl
+import com.example.playlistmaker.search.domain.models.Track
 import com.google.gson.Gson
 
 object Creator {
     private const val TRACK_KEY = "track"
 
-    fun provideInteractor(): PlayerInteractor {
+    fun providePlayerInteractor(): PlayerInteractor {
         return PlayerInteractorImpl(providePlayerRepository())
     }
 
@@ -23,5 +29,14 @@ object Creator {
     fun provideCurrentTrack(intent: Intent): Track {
         val json = intent.getStringExtra(TRACK_KEY)
         return Gson().fromJson(json, Track::class.java)
+    }
+
+
+    fun provideTracksInteractor(context: Context): TracksInteractor {
+        return TracksInteractorImpl(getTracksRepository(context))
+    }
+
+    private fun getTracksRepository(context: Context): TracksRepository {
+        return TracksRepositoryImpl(RetrofitNetworkClient(context))
     }
 }
