--- app/src/main/java/com/example/playlistmaker/domain/models/PlayerState.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.domain.models
-
-enum class PlayerState {
-    DEFAULT,
-    PREPARED,
-    PLAYING,
-    PAUSED
-}
