--- app/src/main/java/com/example/playlistmaker/domain/models/Track.kt
@@ -1,19 +1,0 @@
-package com.example.playlistmaker.domain.models
-
-import com.google.gson.annotations.SerializedName
-
-data class Track (
-    val trackName: String,
-    val artistName: String,
-    @SerializedName("trackTimeMillis")
-    val trackTime: Long,
-    val artworkUrl100: String,
-    val trackId: Int,
-    val collectionName: String,
-    val releaseDate: String,
-    val primaryGenreName: String,
-    val country: String,
-    val previewUrl: String
-) {
-    fun getCoverArtwork() = artworkUrl100.replaceAfterLast('/', "512x512bb.jpg")
-}
