--- app/src/main/java/com/example/playlistmaker/domain/repository/OnPlayerStateChangeListener.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.repository
-
-import com.example.playlistmaker.domain.models.PlayerState
-
-fun interface OnPlayerStateChangeListener {
-    fun onChange(state: PlayerState)
-}
