+++ app/src/main/java/com/example/playlistmaker/main/ui/activity/MainActivity.kt
@@ -1,0 +1,39 @@
+package com.example.playlistmaker.main.ui.activity
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.playlistmaker.media.ui.activity.MediaActivity
+import com.example.playlistmaker.R
+import com.example.playlistmaker.SearchActivity
+import com.example.playlistmaker.SettingsActivity
+import com.google.android.material.button.MaterialButton
+
+class MainActivity : AppCompatActivity() {
+
+    // private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        // binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(R.layout.activity_main)
+
+        val searchButton = findViewById<MaterialButton>(R.id.search_btn)
+        searchButton.setOnClickListener {
+            val searchIntent = Intent(this@MainActivity, SearchActivity::class.java)
+            startActivity(searchIntent)
+        }
+
+        val mediaButton = findViewById<MaterialButton>(R.id.media_btn)
+        mediaButton.setOnClickListener {
+            val mediaIntent = Intent(this@MainActivity, MediaActivity::class.java)
+            startActivity(mediaIntent)
+        }
+
+        val settingsButton = findViewById<MaterialButton>(R.id.settings_btn)
+        settingsButton.setOnClickListener {
+            val settingsIntent = Intent(this@MainActivity, SettingsActivity::class.java)
+            startActivity(settingsIntent)
+        }
+    }
+}
