+++ app/src/main/java/com/example/playlistmaker/media/ui/activity/MediaActivity.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker.media.ui.activity
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.playlistmaker.R
+
+class MediaActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_media)
+    }
+}
