+++ app/src/main/java/com/example/playlistmaker/player/domain/impl/PlayerInteractorImpl.kt
@@ -1,0 +1,39 @@
+package com.example.playlistmaker.player.domain.impl
+
+import com.example.playlistmaker.player.domain.repository.OnPlayerStateChangeListener
+import com.example.playlistmaker.player.domain.models.PlayerState
+import com.example.playlistmaker.player.domain.repository.PlayerRepository
+import com.example.playlistmaker.player.domain.use_case.PlayerInteractor
+
+class PlayerInteractorImpl(
+    private val playerRepository: PlayerRepository
+) : PlayerInteractor {
+
+    override fun preparePlayer(url: String) {
+        playerRepository.preparePlayer(url)
+    }
+
+    override fun startPlayer() {
+        playerRepository.startPlayer()
+    }
+
+    override fun pausePlayer() {
+        playerRepository.pausePlayer()
+    }
+
+    override fun releasePlayer() {
+        playerRepository.releasePlayer()
+    }
+
+    override fun setOnPlayerStateChangeListener(listener: OnPlayerStateChangeListener) {
+        playerRepository.setOnPlayerStateChangeListener(listener)
+    }
+
+    override fun getCurrentPosition(): Int {
+        return playerRepository.getCurrentPosition()
+    }
+
+    override fun getPlayerState(): PlayerState {
+        return playerRepository.getPlayerState()
+    }
+}
