+++ app/src/main/java/com/example/playlistmaker/player/domain/models/PlayerState.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.player.domain.models
+
+enum class PlayerState {
+    DEFAULT,
+    PREPARED,
+    PLAYING,
+    PAUSED
+}
