+++ app/src/main/java/com/example/playlistmaker/player/domain/repository/OnPlayerStateChangeListener.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.player.domain.repository
+
+import com.example.playlistmaker.player.domain.models.PlayerState
+
+fun interface OnPlayerStateChangeListener {
+    fun onChange(state: PlayerState)
+}
