+++ app/src/main/java/com/example/playlistmaker/search/data/dto/Track.kt
@@ -1,0 +1,20 @@
+package com.example.playlistmaker.search.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class TrackDto (
+    val trackName: String,
+    val artistName: String,
+    @SerializedName("trackTimeMillis")
+    val trackTime: Long,
+    val artworkUrl100: String,
+    val trackId: Int,
+    val collectionName: String,
+    val releaseDate: String,
+    val primaryGenreName: String,
+    val country: String,
+    val previewUrl: String
+) {
+    // Убрать из Dto
+    fun getCoverArtwork() = artworkUrl100.replaceAfterLast('/', "512x512bb.jpg")
+}
