+++ app/src/main/java/com/example/playlistmaker/search/data/dto/TrackSearchResponse.kt
@@ -1,0 +1,4 @@
+package com.example.playlistmaker.search.data.dto
+
+data class TrackSearchResponse(val resultCount: Int,
+                               val results: List<TrackDto> ) : Response()
