+++ app/src/main/java/com/example/playlistmaker/search/data/mapper/TrackMapper.kt
@@ -1,0 +1,21 @@
+package com.example.playlistmaker.search.data.mapper
+
+import com.example.playlistmaker.search.data.dto.TrackDto
+import com.example.playlistmaker.search.domain.models.Track
+
+object TrackMapper {
+    fun map(trackDto: TrackDto): Track {
+        return Track(
+            trackName = trackDto.trackName,
+            artistName = trackDto.artistName,
+            trackTime = trackDto.trackTime,
+            artworkUrl100 = trackDto.artworkUrl100,
+            trackId = trackDto.trackId,
+            collectionName = trackDto.collectionName,
+            releaseDate = trackDto.releaseDate.substringBefore("-"),
+            primaryGenreName = trackDto.primaryGenreName,
+            country = trackDto.country,
+            previewUrl = trackDto.previewUrl
+        )
+    }
+}
