+++ app/src/main/java/com/example/playlistmaker/search/data/network/ITunesApiService.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker.search.data.network
+
+import com.example.playlistmaker.search.data.dto.TrackSearchResponse
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ITunesApiService {
+
+    @GET("/search?entity=song")
+    fun searchTracks(@Query("term") text: String) : Call<TrackSearchResponse>
+}
