+++ app/src/main/java/com/example/playlistmaker/search/data/repository/TracksRepositoryImpl.kt
@@ -1,0 +1,29 @@
+package com.example.playlistmaker.search.data.repository
+
+import com.example.playlistmaker.search.data.NetworkClient
+import com.example.playlistmaker.search.data.dto.TrackSearchRequest
+import com.example.playlistmaker.search.data.dto.TrackSearchResponse
+import com.example.playlistmaker.search.data.mapper.TrackMapper
+import com.example.playlistmaker.search.domain.Resource
+import com.example.playlistmaker.search.domain.api.TracksRepository
+import com.example.playlistmaker.search.domain.models.Track
+
+class TracksRepositoryImpl(private val networkClient: NetworkClient) : TracksRepository {
+    override fun searchTracks(expression: String): Resource<List<Track>> {
+        val response = networkClient.doRequest(TrackSearchRequest(expression))
+
+        return when (response.resultCode) {
+            -1 -> Resource.Error("Отсутствует подключение к интернету") // Убрать
+
+            200 -> Resource.Success((response as TrackSearchResponse).results.map {
+                    TrackMapper.map(it)
+                }
+            )
+
+            else -> {
+                Resource.Error("Internal server error") // Убрать
+            }
+        }
+    }
+
+}
