+++ app/src/main/java/com/example/playlistmaker/search/domain/api/TracksInteractor.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker.search.domain.api
+
+import com.example.playlistmaker.search.domain.models.Track
+
+interface TracksInteractor {
+
+    fun searchTracks(expression: String, consumer: TracksConsumer)
+
+    interface TracksConsumer {
+        fun consume(foundTracks: List<Track>?, errorMessage: String?)
+    }
+}
