+++ app/src/main/java/com/example/playlistmaker/search/presentation/TrackInfo.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.search.presentation
+
+data class TrackInfo(
+    val trackName: String,
+    val artistName: String,
+    val trackTime: String, // В строку
+    val artworkUrl100: String
+)
