+++ app/src/main/java/com/example/playlistmaker/settings/data/repository/SettingsRepository.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.settings.data.repository
+
+interface SettingsRepository {
+//    fun getThemeSettings(): ThemeSettings
+//    fun updateThemeSetting(settings: ThemeSettings)
+}
