+++ app/src/main/java/com/example/playlistmaker/settings/domain/interactor/SettingsInteractor.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.settings.domain.interactor
+
+interface SettingsInteractor {
+//    fun getThemeSettings(): ThemeSettings
+//    fun updateThemeSetting(settings: ThemeSettings)
+}
