+++ app/src/main/java/com/example/playlistmaker/settings/ui/view_model/SettingsViewModel.kt
@@ -1,0 +1,9 @@
+package com.example.playlistmaker.settings.ui.view_model
+
+import androidx.lifecycle.ViewModel
+
+//class SettingsViewModel(
+//    private val sharingInteractor: SharingInteractor,
+//    private val settingsInteractor: SettingsInteractor
+//) : ViewModel() {
+//}
