+++ app/src/main/java/com/example/playlistmaker/sharing/domain/interactor/SharingInteractor.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.sharing.domain.interactor
+
+interface SharingInteractor {
+    fun shareApp()
+    fun openTerms()
+    fun openSupport()
+}
